Updates to IStorage, MovieStorage, Movie, and added ListOfStringDeserializer and ListOfStringSerializer.
**IMPORTANT** If we move json files we need to change the code to reflect their new location. **IMPORTANT**

I finished the movie storage class and had to reconfigure the IStorage interface as there were major changes to implementation of MovieStorage. MovieStorage became more of a utility class and has only static methods. So, the interface is basically a clone of Movie Storage. Movie has been updated to add jackson annotations for serializing/deseralizing the bean. Furthermore, I created the ListOfStringSerializer and ListOfStringDeserializer classes that are helpers custom serializers to that help with reading from/writing to json files. Included some extra json files for testing purposes. Not sure if I needed to include those or not. Also included the database as a json file.

 **IMPORTANT** If we move json files we need to change the code to reflect their new location. **IMPORTANT**